refactor: PurchaseInformation 클래스 리팩토링
- 새 파일 PurchaseInformation.js 을 작성 후 내용 옮기기.
- export default 설정 후 View.js 에 import 추가
